+++ NEWS.md
@@ -1,6 +1,8 @@
-# POMADE 0.2.0.9999
+# POMADE 0.2.1
 
 * Added `traffic_light_palette` option to let users specify their own color palette for traffic light strips in the `plot_MADE()` functions.
+* Replaced deprecated `purrr::rerun()` with `purrr::map()`
+* Replaced deprecated `purrr::cross_df()` with `tidyr::expand_grid()`
 
 # POMADE 0.2.0